+++ src/bluesearch/database/article.py
@@ -15,28 +15,19 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
 """Abstraction of scientific article data and related tools."""
+from __future__ import annotations
+
+import html
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
 from pathlib import Path
-from typing import (
-    Dict,
-    Generator,
-    Iterable,
-    List,
-    Optional,
-    Sequence,
-    Tuple,
-    Type,
-    TypeVar,
-    Union,
-)
+from typing import Generator, Iterable, Optional, Sequence, Tuple
 from xml.etree.ElementTree import Element  # nosec
 
 from defusedxml import ElementTree
+from mashumaro import DataClassJSONMixin
 
-# This is for annotating the return value of the Article.parse class method, see
-# https://github.com/python/typing/issues/254#issuecomment-661803922
-_T = TypeVar("_T", bound="Article")
+from bluesearch.database.identifiers import generate_uid
 
 
 class ArticleParser(ABC):
@@ -77,7 +68,7 @@ class ArticleParser(ABC):
 
     @property
     @abstractmethod
-    def paragraphs(self) -> Iterable[Tuple[str, str]]:
+    def paragraphs(self) -> Iterable[tuple[str, str]]:
         """Get all paragraphs and titles of sections they are part of.
 
         Returns
@@ -87,9 +78,54 @@ class ArticleParser(ABC):
             is the section title, the second the paragraph content.
         """
 
+    @property
+    def pubmed_id(self) -> Optional[str]:
+        """Get Pubmed ID.
 
-class PubmedXMLParser(ArticleParser):
-    """Parser for PubMed XML files using the JATS Journal Publishing DTD.
+        Returns
+        -------
+        str or None
+            Pubmed ID if specified, otherwise None.
+        """
+        return None
+
+    @property
+    def pmc_id(self) -> Optional[str]:
+        """Get PMC ID.
+
+        Returns
+        -------
+        str or None
+            PMC ID if specified, otherwise None.
+        """
+        return None
+
+    @property
+    def doi(self) -> Optional[str]:
+        """Get DOI.
+
+        Returns
+        -------
+        str or None
+            DOI if specified, otherwise None.
+        """
+        return None
+
+    @property
+    def uid(self) -> Optional[str]:
+        """Generate unique ID of the article based on different identifiers.
+
+        Returns
+        -------
+        str or None
+            If at least one identifier exists, unique id is created. Otherwise,
+            the returned uid is None.
+        """
+        return generate_uid((self.pubmed_id, self.pmc_id, self.doi))
+
+
+class PMCXMLParser(ArticleParser):
+    """Parser for PubMed Central XML files using the JATS Journal Publishing DTD.
 
     Parameters
     ----------
@@ -97,9 +133,10 @@ class PubmedXMLParser(ArticleParser):
         The path to the XML file from PubMed.
     """
 
-    def __init__(self, path: Union[str, Path]) -> None:
+    def __init__(self, path: str | Path) -> None:
         super().__init__()
         self.content = ElementTree.parse(str(path))
+        self.ids = self.get_ids()
 
     @property
     def title(self) -> str:
@@ -111,7 +148,7 @@ class PubmedXMLParser(ArticleParser):
             The article title.
         """
         titles = self.content.find(".//title-group/article-title")
-        return self.text_content(titles)
+        return self._element_to_str(titles)
 
     @property
     def authors(self) -> Generator[str, None, None]:
@@ -126,8 +163,8 @@ class PubmedXMLParser(ArticleParser):
             ".//contrib-group/contrib[@contrib-type='author']"
         )
         for author in authors:
-            given_names = self.text_content(author.find("name/given-names"))
-            surname = self.text_content(author.find("name/surname"))
+            given_names = self._element_to_str(author.find("name/given-names"))
+            surname = self._element_to_str(author.find("name/surname"))
             if given_names == "" or surname == "":
                 # In rare cases, an author may not have a given name or a surname,
                 # e.g. it could be an organization. We decide to skip those.
@@ -146,10 +183,10 @@ class PubmedXMLParser(ArticleParser):
         """
         abstract_pars = self.content.findall(".//abstract//p")
         for paragraph in abstract_pars:
-            yield self.text_content(paragraph)
+            yield self._element_to_str(paragraph)
 
     @property
-    def paragraphs(self) -> Generator[Tuple[str, str], None, None]:
+    def paragraphs(self) -> Generator[tuple[str, str], None, None]:
         """Get all paragraphs and titles of sections they are part of.
 
         Paragraphs can be parts of text body, or figure or table captions.
@@ -175,7 +212,7 @@ class PubmedXMLParser(ArticleParser):
         section_dirs = self.get_paragraphs_sections_mapping()
         for paragraph in self.content.findall(".//p"):
             if paragraph not in exclude_list:
-                text = self.text_content(paragraph)
+                text = self._element_to_str(paragraph)
                 section_title = ""
                 if paragraph in section_dirs:
                     section_title = section_dirs[paragraph]
@@ -187,7 +224,7 @@ class PubmedXMLParser(ArticleParser):
             fig_captions = fig.findall("caption")
             if fig_captions is None:
                 continue
-            caption = " ".join(self.text_content(c) for c in list(fig_captions))
+            caption = " ".join(self._element_to_str(c) for c in list(fig_captions))
             yield "Figure Caption", caption
 
         # Table captions
@@ -198,10 +235,63 @@ class PubmedXMLParser(ArticleParser):
             )
             if caption_elements is None:
                 continue
-            caption = " ".join(self.text_content(c) for c in caption_elements)
+            caption = " ".join(self._element_to_str(c) for c in caption_elements)
             yield "Table Caption", caption
 
-    def get_paragraphs_sections_mapping(self) -> Dict[Element, str]:
+    @property
+    def pubmed_id(self) -> Optional[str]:
+        """Get Pubmed ID.
+
+        Returns
+        -------
+        str or None
+            Pubmed ID if specified, otherwise None.
+        """
+        return self.ids.get("pmid")
+
+    @property
+    def pmc_id(self) -> Optional[str]:
+        """Get PMC ID.
+
+        Returns
+        -------
+        str or None
+            PMC ID if specified, otherwise None.
+        """
+        return self.ids.get("pmc")
+
+    @property
+    def doi(self) -> Optional[str]:
+        """Get DOI.
+
+        Returns
+        -------
+        str or None
+            DOI if specified, otherwise None.
+        """
+        return self.ids.get("doi")
+
+    def get_ids(self) -> dict[str, str]:
+        """Get all specified IDs of the paper.
+
+        Returns
+        -------
+        ids : dict
+            Dictionary whose keys are ids type and value are ids values.
+        """
+        ids = {}
+        article_ids = self.content.findall("./front/article-meta/article-id")
+
+        for article_id in article_ids:
+
+            if "pub-id-type" not in article_id.attrib.keys():
+                continue
+
+            ids[article_id.attrib["pub-id-type"]] = article_id.text
+
+        return ids
+
+    def get_paragraphs_sections_mapping(self) -> dict[Element, str]:
         """Construct mapping between all paragraphs and their section name.
 
         Returns
@@ -218,7 +308,7 @@ class PubmedXMLParser(ArticleParser):
             for element in sec:
 
                 if element.tag == "title":
-                    section_title = self.text_content(element)
+                    section_title = self._element_to_str(element)
                 elif element.tag == "p":
                     section_paragraphs.append(element)
 
@@ -227,24 +317,210 @@ class PubmedXMLParser(ArticleParser):
 
         return mapping
 
-    @staticmethod
-    def text_content(element: Optional[Element]) -> str:
-        """Extract all text of an element and of its descendants (at any depth).
+    def _inner_text(self, element: Element) -> str:
+        """Convert all inner text and sub-elements to one string.
+
+        In short, we collect all the inner text while also converting all
+        sub-elements that we encounter to strings using ``self._element_to_str``.
+        All escaped HTML in the raw text is unescaped.
+
+        For example, if schematically the element is given by
+
+            element = "<p>I <bold>like</bold> python &amp; ice cream.<p>"
+
+        then ``_inner_text(element)`` would give
+
+            "I like python & ice cream."
+
+        provided that "<bold>like</bold>" is resolved to "like" by the
+        ``self._element_to_str`` method.
 
         Parameters
         ----------
-        element : etree._Element or None
-            XML element to parse.
+        element
+            The input XML element.
 
         Returns
         -------
         str
-            Entire text of the element and its descendants.
+            The inner text and sub-elements converted to one single string.
+        """
+        text_parts = [html.unescape(element.text or "")]
+        for sub_element in element:
+            # recursively parse the sub-element
+            text_parts.append(self._element_to_str(sub_element))
+            # don't forget the text after the sub-element
+            text_parts.append(html.unescape(sub_element.tail or ""))
+        return "".join(text_parts).strip()
+
+    def _element_to_str(self, element: Element | None) -> str:
+        """Convert an element and all its contents to a string.
+
+        Parameters
+        ----------
+        element
+            The input XML element.
+
+        Returns
+        -------
+        str
+            A parsed string representation of the input XML element.
         """
         if element is None:
             return ""
+
+        if element.tag in {
+            "bold",
+            "italic",
+            "monospace",
+            "p",
+            "sc",
+            "styled-content",
+            "underline",
+            "xref",
+        }:
+            # Mostly styling tags for which getting the inner text is enough.
+            # Currently this is the same as the default handling. Writing it out
+            # explicitly here to decouple from the default handling, which may
+            # change in the future.
+            return self._inner_text(element)
+        elif element.tag == "sub":
+            return f"_{self._inner_text(element)}"
+        elif element.tag == "sup":
+            return f"^{self._inner_text(element)}"
+        elif element.tag == "inline-formula":
+            return "FORMULA"
+        elif element.tag == "disp-formula":
+            return "\nFORMULA-BLOCK"
+        elif element.tag in {"ext-link", "uri"}:
+            return "URL"
+        elif element.tag == "email":
+            return "EMAIL"
         else:
-            return "".join(element.itertext()).strip()
+            # Default handling for all other element tags
+            return self._inner_text(element)
+
+
+class PubMedXMLParser(ArticleParser):
+    """Parser for PubMed abstract."""
+
+    def __init__(self, data: Element | Path | str) -> None:
+        super().__init__()
+        self.content: ElementTree
+        if isinstance(data, Element):
+            self.content = data
+        else:
+            self.content = ElementTree.parse(str(data))
+
+    @property
+    def title(self) -> str:
+        """Get the article title.
+
+        Returns
+        -------
+        str
+            The article title.
+        """
+        title = self.content.find("./MedlineCitation/Article/ArticleTitle")
+        return title.text
+
+    @property
+    def authors(self) -> Iterable[str]:
+        """Get all author names.
+
+        Returns
+        -------
+        iterable of str
+            All authors.
+        """
+        authors = self.content.find("./MedlineCitation/Article/AuthorList")
+
+        if authors is None:
+            # No author to parse: stop and return an empty iterable.
+            return ()
+
+        for author in authors:
+            # Author entries with 'ValidYN' == 'N' are incorrect entries:
+            # https://dtd.nlm.nih.gov/ncbi/pubmed/doc/out/190101/att-ValidYN.html.
+            if author.get("ValidYN") == "Y":
+                # 'LastName' is a required field if there is no 'CollectiveName'.
+                lastname = author.find("LastName")
+                # 'ForeName' is an optional field only used with 'LastName'.
+                forenames = author.find("ForeName")
+
+                parts = (forenames, lastname)
+                name = [x.text for x in parts if x is not None]
+                if len(name) > 0:
+                    yield " ".join(name)
+
+    @property
+    def abstract(self) -> Iterable[str]:
+        """Get a sequence of paragraphs in the article abstract.
+
+        Returns
+        -------
+        iterable of str
+            The paragraphs of the article abstract.
+        """
+        paragraphs = self.content.find("./MedlineCitation/Article/Abstract")
+
+        if paragraphs is None:
+            # No paragraphs to parse: stop and return an empty iterable.
+            return ()
+
+        for paragraph in paragraphs.iter("AbstractText"):
+            yield paragraph.text
+
+    @property
+    def paragraphs(self) -> Iterable[tuple[str, str]]:
+        """Get all paragraphs and titles of sections they are part of.
+
+        Returns
+        -------
+        iterable of (str, str)
+            For each paragraph a tuple with two strings is returned. The first
+            is the section title, the second the paragraph content.
+        """
+        # No paragraph to parse in PubMed article sets: return an empty iterable.
+        return ()
+
+    @property
+    def pubmed_id(self) -> Optional[str]:
+        """Get Pubmed ID.
+
+        Returns
+        -------
+        str or None
+            Pubmed ID if specified, otherwise None.
+        """
+        pubmed_id = self.content.find("./MedlineCitation/PMID")
+        return pubmed_id.text
+
+    @property
+    def pmc_id(self) -> Optional[str]:
+        """Get PMC ID.
+
+        Returns
+        -------
+        str or None
+            PMC ID if specified, otherwise None.
+        """
+        pmc_id = self.content.find(
+            "./PubmedData/ArticleIdList/ArticleId[@IdType='pmc']"
+        )
+        return None if pmc_id is None else pmc_id.text
+
+    @property
+    def doi(self) -> Optional[str]:
+        """Get DOI.
+
+        Returns
+        -------
+        str or None
+            DOI if specified, otherwise None.
+        """
+        doi = self.content.find("./PubmedData/ArticleIdList/ArticleId[@IdType='doi']")
+        return None if doi is None else doi.text
 
 
 class CORD19ArticleParser(ArticleParser):
@@ -311,7 +587,7 @@ class CORD19ArticleParser(ArticleParser):
             yield author_str
 
     @property
-    def abstract(self) -> List[str]:
+    def abstract(self) -> list[str]:
         """Get a sequence of paragraphs in the article abstract.
 
         Returns
@@ -325,7 +601,7 @@ class CORD19ArticleParser(ArticleParser):
         return [paragraph["text"] for paragraph in self.data["abstract"]]
 
     @property
-    def paragraphs(self) -> Generator[Tuple[str, str], None, None]:
+    def paragraphs(self) -> Generator[tuple[str, str], None, None]:
         """Get all paragraphs and titles of sections they are part of.
 
         Yields
@@ -341,22 +617,37 @@ class CORD19ArticleParser(ArticleParser):
         for ref_entry in self.data["ref_entries"].values():
             yield "Caption", ref_entry["text"]
 
+    @property
+    def pmc_id(self) -> Optional[str]:
+        """Get PMC ID.
+
+        Returns
+        -------
+        str or None
+            PMC ID if specified, otherwise None.
+        """
+        return self.data.get("paper_id")
+
     def __str__(self):
         """Get the string representation of the parser instance."""
         return f'CORD-19 article ID={self.data["paper_id"]}'
 
 
 @dataclass(frozen=True)
-class Article:
+class Article(DataClassJSONMixin):
     """Abstraction of a scientific article and its contents."""
 
     title: str
     authors: Sequence[str]
     abstract: Sequence[str]
     section_paragraphs: Sequence[Tuple[str, str]]
+    pubmed_id: Optional[str] = None
+    pmc_id: Optional[str] = None
+    doi: Optional[str] = None
+    uid: Optional[str] = None
 
     @classmethod
-    def parse(cls: Type[_T], parser: ArticleParser) -> _T:
+    def parse(cls, parser: ArticleParser) -> Article:
         """Parse an article through a parser.
 
         Parameters
@@ -368,12 +659,18 @@ class Article:
         authors = tuple(parser.authors)
         abstract = tuple(parser.abstract)
         section_paragraphs = tuple(parser.paragraphs)
+        pubmed_id = parser.pubmed_id
+        pmc_id = parser.pmc_id
+        doi = parser.doi
+        uid = parser.uid
 
-        return cls(title, authors, abstract, section_paragraphs)
+        return cls(
+            title, authors, abstract, section_paragraphs, pubmed_id, pmc_id, doi, uid
+        )
 
     def iter_paragraphs(
         self, with_abstract: bool = False
-    ) -> Generator[Tuple[str, str], None, None]:
+    ) -> Generator[tuple[str, str], None, None]:
         """Iterate over all paragraphs in the article.
 
         Parameters